Move getMatch into selectors
